requirements.txt
# Name: Terence Anquandah
# Index Number: 10022200077

# =========================
# CORE APP (UI)
# =========================
streamlit>=1.32,<2
python-dotenv>=1.0,<2

# =========================
# FASTAPI BACKEND
# =========================
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
pydantic>=2.0,<3

# =========================
# DATA PROCESSING
# =========================
pandas>=2.0,<3
numpy>=1.24,<3

# =========================
# PDF PROCESSING
# =========================
# Include both engines for robustness; select via config.
pdfplumber>=0.11,<1
PyMuPDF>=1.23,<2

# =========================
# EMBEDDINGS (RAG CORE)
# =========================
sentence-transformers>=2.7,<4
# Explicit torch dependency improves install reliability on some environments.
torch

# =========================
# VECTOR SEARCH (FAISS)
# =========================
faiss-cpu>=1.8,<2

# =========================
# KEYWORD SEARCH (TF-IDF)
# =========================
scikit-learn>=1.3,<2

# =========================
# UTILITIES
# =========================
tqdm>=4.66,<5
rich>=13,<14

# =========================
# LLM API
# =========================
google-cloud-aiplatform